+++ unliftio-core/ChangeLog.md
@@ -2,6 +2,7 @@
 
 * Doc improvements.
 * Inline functions in `Control.Monad.IO.Unlift` module [#4](https://github.com/fpco/unliftio/pull/4).
+* Fully polymorphic `withRunInIO` [#12](https://github.com/fpco/unliftio/pull/12).
 
 ## 0.1.0.0